backend/generate-report/requirements.txt
psycopg2-binary>=2.9.9
reportlab>=4.0.7
boto3>=1.34.0
reportlab-qrcode>=0.1.4